requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
boto3==1.34.0
python-multipart==0.0.6
PyJWT==2.8.0
python-jose[cryptography]==3.3.0